Why:
Some generative art collections have inscribed all art within a single
parent inscription to save considerable storage space on the
blockchain.
The ord explorer is unable to currently support collections inscribed
in this manner. This commit adds this support.
What:

In order to support these generative art collections, we need to be able
to inscribe the metadata for children in an agreed upon JSON spec.
These child inscriptions are JSON documents containing url parameters
that instruct the parent inscription to
render a specific item in the collection.
How:

Since all art is contained in the parent, we added a simple redirect to
render the parent inscription with included url parameters.
There are 3 constraints which must be in the child inscription before
the redirect can happen:
The child inscription must have a parent inscription.
The child inscription's body data must be valid JSON.
The child inscription's body JSON data must contain a field named
"use_p", with any value.
Proposed JSON standard:

```
{
   "use_p": 1,
   "params": ["tokenID=4969"]
}
```

Co-authored-by: clrke <clrke@metagood.com>